+++ index.js
@@ -11,11 +11,8 @@ import './src/pattern.class'; // optional pattern
 import './src/shadow.class'; // optional shadow
 import './src/canvas/static_canvas.class';
 import './src/canvas/canvas_events'; // optional interaction
-import './src/mixins/canvas_dataurl_exporter.mixin';
-import './src/mixins/canvas_serialization.mixin'; // optional serialization
 import './src/canvas/canvas_gestures.mixin'; // optional gestures
 import './src/mixins/canvas_animation.mixin'; // optional animation
-import './src/mixins/canvas_straightening.mixin'; // optional animation
 import './src/shapes/Object/FabricObject';
 import './src/mixins/object_stacking.mixin'; // removed in #8461
 import './src/mixins/stateful.mixin'; // will die soon