+++ src/canvas/static_canvas.class.ts
@@ -12,11 +12,14 @@ import { Point } from '../point.class';
 import type { FabricObject } from '../shapes/Object/FabricObject';
 import { TCachedFabricObject } from '../shapes/Object/Object';
 import { Rect } from '../shapes/rect.class';
-import type {
+import {
+  ImageFormat,
   TCornerPoint,
+  TDataUrlOptions,
   TFiller,
   TMat2D,
   TSize,
+  TToCanvasElementOptions,
   TValidToObjectMethod,
 } from '../typedefs';
 import { cancelAnimFrame, requestAnimFrame } from '../util/animate';
@@ -27,8 +30,13 @@ import {
 } from '../util/dom_misc';
 import { removeFromArray } from '../util/internals';
 import { uid } from '../util/internals/uid';
-import { createCanvasElement, isHTMLCanvas } from '../util/misc/dom';
+import { createCanvasElement, isHTMLCanvas, toDataURL } from '../util/misc/dom';
 import { invertTransform, transformPoint } from '../util/misc/matrix';
+import {
+  enlivenObjectEnlivables,
+  EnlivenObjectOptions,
+  enlivenObjects,
+} from '../util/misc/objectEnlive';
 import { pick } from '../util/misc/pick';
 import { matrixToSVG } from '../util/misc/svgParsing';
 import { toFixed } from '../util/misc/toFixed';
@@ -58,6 +66,10 @@ export type TSVGExportOptions = {
   width?: string;
   height?: string;
   reviver?: TSVGReviver;
+};
+
+type TCanvasHydrationOption = {
+  signal?: AbortSignal;
 };
 
 /**
@@ -273,7 +285,9 @@ export class StaticCanvas<
   _offset: { left: number; top: number };
   protected hasLostContext: boolean;
   protected nextRenderHandle: number;
-  protected __cleanupTask: {
+
+  // reference to
+  protected __cleanupTask?: {
     (): void;
     kill: (reason?: any) => void;
   };
@@ -456,7 +470,7 @@ export class StaticCanvas<
     }
     this.lowerCanvasEl.classList.add('lower-canvas');
     this.lowerCanvasEl.setAttribute('data-fabric', 'main');
-    this.contextContainer = this.lowerCanvasEl.getContext('2d');
+    this.contextContainer = this.lowerCanvasEl.getContext('2d')!;
   }
 
   /**
@@ -1061,7 +1075,7 @@ export class StaticCanvas<
         : null;
     return {
       version: VERSION,
-      ...pick(this, propertiesToInclude),
+      ...pick(this, propertiesToInclude as (keyof this)[]),
       objects: this._objects
         .filter((object) => !object.excludeFromExport)
         .map((instance) =>
@@ -1294,7 +1308,7 @@ export class StaticCanvas<
    * @return {String}
    */
   createSVGRefElementsMarkup(): string {
-    return ['background', 'overlay']
+    return (['background', 'overlay'] as const)
       .map((prop) => {
         const fill = this[`${prop}Color`];
         if (isFiller(fill)) {
@@ -1644,6 +1658,217 @@ export class StaticCanvas<
   }
 
   /**
+   * Populates canvas with data from the specified JSON.
+   * JSON format must conform to the one of {@link fabric.Canvas#toJSON}
+   *
+   * **IMPORTANT**: It is recommended to abort loading tasks before calling this method to prevent race conditions and unnecessary networking
+   *
+   * @param {String|Object} json JSON string or object
+   * @param {Function} [reviver] Method for further parsing of JSON elements, called after each fabric object created.
+   * @param {Object} [options] options
+   * @param {AbortSignal} [options.signal] see https://developer.mozilla.org/en-US/docs/Web/API/AbortController/signal
+   * @return {Promise<Canvas | StaticCanvas>} instance
+   * @tutorial {@link http://fabricjs.com/fabric-intro-part-3#deserialization}
+   * @see {@link http://jsfiddle.net/fabricjs/fmgXt/|jsFiddle demo}
+   * @example <caption>loadFromJSON</caption>
+   * canvas.loadFromJSON(json).then((canvas) => canvas.requestRenderAll());
+   * @example <caption>loadFromJSON with reviver</caption>
+   * canvas.loadFromJSON(json, function(o, object) {
+   *   // `o` = json object
+   *   // `object` = fabric.Object instance
+   *   // ... do some stuff ...
+   * }).then((canvas) => {
+   *   ... canvas is restored, add your code.
+   * });
+   *
+   */
+  loadFromJSON(
+    json: string | Record<string, any>,
+    reviver?: EnlivenObjectOptions['reviver'],
+    { signal }: TCanvasHydrationOption = {}
+  ): Promise<this> {
+    if (!json) {
+      return Promise.reject(new Error('fabric.js: `json` is undefined'));
+    }
+
+    // parse json if it wasn't already
+    const serialized = typeof json === 'string' ? JSON.parse(json) : json;
+    const {
+      objects = [],
+      backgroundImage,
+      background,
+      overlayImage,
+      overlay,
+      clipPath,
+    } = serialized;
+    const renderOnAddRemove = this.renderOnAddRemove;
+    this.renderOnAddRemove = false;
+
+    return Promise.all([
+      enlivenObjects(objects, {
+        reviver,
+        signal,
+      }),
+      enlivenObjectEnlivables(
+        {
+          backgroundImage,
+          backgroundColor: background,
+          overlayImage,
+          overlayColor: overlay,
+          clipPath,
+        },
+        { signal }
+      ),
+    ]).then(([enlived, enlivedMap]) => {
+      this.clear();
+      this.add(...enlived);
+      this.set(serialized);
+      this.set(enlivedMap);
+      this.renderOnAddRemove = renderOnAddRemove;
+      return this;
+    });
+  }
+
+  /**
+   * Clones canvas instance
+   * @param {string[]} [properties] Array of properties to include in the cloned canvas and children
+   * @returns {Promise<Canvas | StaticCanvas>}
+   */
+  clone(properties: string[]): Promise<this> {
+    const data = this.toObject(properties);
+    const canvas = this.cloneWithoutData();
+    // @ts-ignore
+    return canvas.loadFromJSON(data);
+  }
+
+  /**
+   * Clones canvas instance without cloning existing data.
+   * This essentially copies canvas dimensions since loadFromJSON does not affect canvas size.
+   * @returns {StaticCanvas}
+   */
+  cloneWithoutData(): StaticCanvas {
+    const el = createCanvasElement();
+    el.width = this.width;
+    el.height = this.height;
+    // this seems wrong. either Canvas or StaticCanvas
+    return new StaticCanvas(el);
+  }
+
+  /**
+   * Exports canvas element to a dataurl image. Note that when multiplier is used, cropping is scaled appropriately
+   * @param {Object} [options] Options object
+   * @param {String} [options.format=png] The format of the output image. Either "jpeg" or "png"
+   * @param {Number} [options.quality=1] Quality level (0..1). Only used for jpeg.
+   * @param {Number} [options.multiplier=1] Multiplier to scale by, to have consistent
+   * @param {Number} [options.left] Cropping left offset. Introduced in v1.2.14
+   * @param {Number} [options.top] Cropping top offset. Introduced in v1.2.14
+   * @param {Number} [options.width] Cropping width. Introduced in v1.2.14
+   * @param {Number} [options.height] Cropping height. Introduced in v1.2.14
+   * @param {Boolean} [options.enableRetinaScaling] Enable retina scaling for clone image. Introduce in 2.0.0
+   * @param {(object: fabric.Object) => boolean} [options.filter] Function to filter objects.
+   * @return {String} Returns a data: URL containing a representation of the object in the format specified by options.format
+   * @see {@link https://jsfiddle.net/xsjua1rd/ demo}
+   * @example <caption>Generate jpeg dataURL with lower quality</caption>
+   * var dataURL = canvas.toDataURL({
+   *   format: 'jpeg',
+   *   quality: 0.8
+   * });
+   * @example <caption>Generate cropped png dataURL (clipping of canvas)</caption>
+   * var dataURL = canvas.toDataURL({
+   *   format: 'png',
+   *   left: 100,
+   *   top: 100,
+   *   width: 200,
+   *   height: 200
+   * });
+   * @example <caption>Generate double scaled png dataURL</caption>
+   * var dataURL = canvas.toDataURL({
+   *   format: 'png',
+   *   multiplier: 2
+   * });
+   * @example <caption>Generate dataURL with objects that overlap a specified object</caption>
+   * var myObject;
+   * var dataURL = canvas.toDataURL({
+   *   filter: (object) => object.isContainedWithinObject(myObject) || object.intersectsWithObject(myObject)
+   * });
+   */
+  toDataURL(options = {} as TDataUrlOptions): string {
+    const {
+      format = ImageFormat.png,
+      quality = 1,
+      multiplier = 1,
+      enableRetinaScaling = false,
+    } = options;
+    const finalMultiplier =
+      multiplier * (enableRetinaScaling ? this.getRetinaScaling() : 1);
+
+    return toDataURL(
+      this.toCanvasElement(finalMultiplier, options),
+      format,
+      quality
+    );
+  }
+
+  /**
+   * Create a new HTMLCanvas element painted with the current canvas content.
+   * No need to resize the actual one or repaint it.
+   * Will transfer object ownership to a new canvas, paint it, and set everything back.
+   * This is an intermediary step used to get to a dataUrl but also it is useful to
+   * create quick image copies of a canvas without passing for the dataUrl string
+   * @param {Number} [multiplier] a zoom factor.
+   * @param {Object} [options] Cropping informations
+   * @param {Number} [options.left] Cropping left offset.
+   * @param {Number} [options.top] Cropping top offset.
+   * @param {Number} [options.width] Cropping width.
+   * @param {Number} [options.height] Cropping height.
+   * @param {(object: fabric.Object) => boolean} [options.filter] Function to filter objects.
+   */
+  toCanvasElement(
+    multiplier = 1,
+    { width, height, left, top, filter } = {} as TToCanvasElementOptions
+  ): HTMLCanvasElement {
+    const scaledWidth = (width || this.width) * multiplier,
+      scaledHeight = (height || this.height) * multiplier,
+      zoom = this.getZoom(),
+      originalWidth = this.width,
+      originalHeight = this.height,
+      newZoom = zoom * multiplier,
+      vp = this.viewportTransform,
+      translateX = (vp[4] - (left || 0)) * multiplier,
+      translateY = (vp[5] - (top || 0)) * multiplier,
+      // @ts-ignore
+      originalInteractive = this.interactive,
+      newVp = [newZoom, 0, 0, newZoom, translateX, translateY] as TMat2D,
+      originalRetina = this.enableRetinaScaling,
+      canvasEl = createCanvasElement(),
+      // @ts-ignore
+      originalContextTop = this.contextTop,
+      objectsToRender = filter ? this._objects.filter(filter) : this._objects;
+    canvasEl.width = scaledWidth;
+    canvasEl.height = scaledHeight;
+    // @ts-ignore
+    this.contextTop = null;
+    this.enableRetinaScaling = false;
+    // @ts-ignore
+    this.interactive = false;
+    this.viewportTransform = newVp;
+    this.width = scaledWidth;
+    this.height = scaledHeight;
+    this.calcViewportBoundaries();
+    this.renderCanvas(canvasEl.getContext('2d')!, objectsToRender);
+    this.viewportTransform = vp;
+    this.width = originalWidth;
+    this.height = originalHeight;
+    this.calcViewportBoundaries();
+    // @ts-ignore
+    this.interactive = originalInteractive;
+    this.enableRetinaScaling = originalRetina;
+    // @ts-ignore
+    this.contextTop = originalContextTop;
+    return canvasEl;
+  }
+
+  /**
    * Waits until rendering has settled to destroy the canvas
    * @returns {Promise<boolean>} a promise resolving to `true` once the canvas has been destroyed or to `false` if the canvas has was already destroyed
    * @throws if aborted by a consequent call
@@ -1695,7 +1920,6 @@ export class StaticCanvas<
       this.overlayImage.dispose();
     }
     this.overlayImage = null;
-    this._iTextInstances = null;
     // @ts-expect-error disposing
     this.contextContainer = null;
     const canvasElement = this.lowerCanvasEl;
@@ -1723,39 +1947,41 @@ export class StaticCanvas<
   }
 }
 
-Object.assign(
-  StaticCanvas.prototype,
-  {
-    backgroundColor: '',
-    backgroundImage: null,
-    overlayColor: '',
-    overlayImage: null,
-    includeDefaultValues: true,
-    stateful: false,
-    renderOnAddRemove: true,
-    controlsAboveOverlay: false,
-    allowTouchScrolling: false,
-    imageSmoothingEnabled: true,
-    viewportTransform: iMatrix.concat(),
-    backgroundVpt: true,
-    overlayVpt: true,
-    enableRetinaScaling: true,
-    svgViewportTransformation: true,
-    skipOffscreen: true,
-    clipPath: undefined,
-  },
-  fabric.DataURLExporter
-);
+Object.assign(StaticCanvas.prototype, {
+  backgroundColor: '',
+  backgroundImage: null,
+  overlayColor: '',
+  overlayImage: null,
+  includeDefaultValues: true,
+  stateful: false,
+  renderOnAddRemove: true,
+  controlsAboveOverlay: false,
+  allowTouchScrolling: false,
+  imageSmoothingEnabled: true,
+  viewportTransform: iMatrix.concat(),
+  backgroundVpt: true,
+  overlayVpt: true,
+  enableRetinaScaling: true,
+  svgViewportTransformation: true,
+  skipOffscreen: true,
+  clipPath: undefined,
+});
 
 if (fabric.isLikelyNode) {
-  StaticCanvas.prototype.createPNGStream = function () {
-    const impl = getNodeCanvas(this.lowerCanvasEl);
-    return impl && impl.createPNGStream();
-  };
-  StaticCanvas.prototype.createJPEGStream = function (opts: any) {
-    const impl = getNodeCanvas(this.lowerCanvasEl);
-    return impl && impl.createJPEGStream(opts);
-  };
+  Object.assign(StaticCanvas.prototype, {
+    createPNGStream() {
+      const impl = getNodeCanvas(
+        (this as unknown as StaticCanvas).lowerCanvasEl
+      );
+      return impl && impl.createPNGStream();
+    },
+    createJPEGStream(opts: any) {
+      const impl = getNodeCanvas(
+        (this as unknown as StaticCanvas).lowerCanvasEl
+      );
+      return impl && impl.createJPEGStream(opts);
+    },
+  });
 }
 
 fabric.StaticCanvas = StaticCanvas;